feat: use XDG base directory structure in Linux
The Linux version now uses the XDG Base Directory Structure, so that
instead of having everything in `~/.gaiasky`, we now have
`~/.local/share/gaiasky` for the data files (scripts, screenshots, data
folder, etc) and `~/.config/gaiasky` for the controller mappings and
config file. You can move the files to the new locations and everything
should just work.
Windows and macOS versions are unaffected by this commit.