feat: add extra attributes, colormaps for highlighting
Add color maps for highlighting, additionally to the current plain
color. When highlighting a dataset, an arbitrary attribute can be
selected, anlong with a minimum and maximum mapping values, in order to
be displayed using a color map.
Loading datasets from VOTables will also load extra (numeric) attributes
and make them available to the filtering and colormap highlighting
modules.