build: use externally built JDK for appimage
The JDK in Arch Linux repositories is built against glibc 2.33,
which is cutting edge and not available in most non-rolling distros.
This caused the appimage to not work in these distros. This commit
changes the build system so that an externally built JDK is downloaded,
unpacked and used to build the packaged JRE with `jlink`. We use
the JDK provided by AdoptOpenJDK.

Fixes #361